+++ src/main/java/com/tmdigital/gestiondestock/services/impl/CompanyServiceImpl.java
@@ -7,7 +7,9 @@ import org.springframework.stereotype.Service;
 import com.tmdigital.gestiondestock.dto.CompanyDto;
 import com.tmdigital.gestiondestock.exception.ErrorCodes;
 import com.tmdigital.gestiondestock.exception.InvalidEntityException;
+import com.tmdigital.gestiondestock.model.User;
 import com.tmdigital.gestiondestock.repository.CompanyRepository;
+import com.tmdigital.gestiondestock.repository.UserRepository;
 import com.tmdigital.gestiondestock.services.CompanyService;
 import com.tmdigital.gestiondestock.validator.CompanyValidator;
 
@@ -17,9 +19,11 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class CompanyServiceImpl implements CompanyService {
     private CompanyRepository companyRepository;
+    private UserRepository userRepository;
 
-    public CompanyServiceImpl(CompanyRepository companyRepository) {
+    public CompanyServiceImpl(CompanyRepository companyRepository, UserRepository userRepository) {
         this.companyRepository = companyRepository;
+        this.userRepository = userRepository;
     }
 
     @Override
@@ -60,6 +64,14 @@ public class CompanyServiceImpl implements CompanyService {
             log.error("L'identifiant est nul");
             return;
         }
+
+        List<User> users = userRepository.findAllByCompanyId(id);
+
+        if (!users.isEmpty()) {
+            log.error("Impossible de supprimer l'entreprise");
+            throw new InvalidEntityException("Impossible de supprimer l'entreprise avec l'id = " + id, ErrorCodes.ARTICLE_ALREADY_IN_USE);
+        }
+
         companyRepository.deleteById(id);
     }
